requirements.txt
azure-functions
azure-cosmosdb-table
beautifulsoup4
requests
datetime
sendgrid
jsonpickle